+++ .gitignore
@@ -37,3 +37,15 @@ gradle-app.setting
 
 # VSCode
 .vscode
+
+# OS generated files #
+######################
+*/.DS_Store
+.DS_Store
+.DS_Store?
+._*
+.Spotlight-V100
+.Trashes
+ehthumbs.db
+Thumbs.db
+
