+++ src/main/java/com/nrh/api/module/nr/APIKeyset.java
@@ -1,4 +1,4 @@
-package com.adg.newrelic.api;
+package com.nrh.api.module.nr;
 
 import com.typesafe.config.Config;
