+++ src/main/java/com/nrh/api/module/nr/task/ExampleTask.java
@@ -1,4 +1,4 @@
-package com.adg.newrelic.copier.task;
+package com.nrh.api.module.nr.task;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
