+++ src/test/java/com/nrh/api/TestInsights.java
@@ -1,4 +1,4 @@
-package com.adg.newrelic.api;
+package com.nrh.api;
 
 import static org.junit.Assert.*;
 
@@ -11,6 +11,8 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.nrh.api.module.nr.APIKeyset;
+import com.nrh.api.module.nr.Insights;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
